Crate lang

Language handling

This contains all the keywords currently available for fck, as well as the functions that allow the extended use of them. This crate (in the fck repo) also includes the fckl equivalents folder that has the .fckl language files equivalent to all the currently included languages

Modules

de

Language file for German

en

Language file for English

fr

Language file for French

keywords

Structs for the languages

ko

Language file for Korean

Functions

get_associated_keywords

Returns (assuming the language code is valid) the keywords relating to the given language code

get_associated_messages

Returns (assuming the language code is valid) the messages relating to the given language code